requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pika==1.2.0
python-dotenv==0.17.0
requests==2.25.1
urllib3==1.26.5
# confluent-kafka==1.7.0
lakefs_client==0.80.1
# coverage==6.5.0
twilio==7.16.2

